rethinking and updating
