+++ .prettierrc
@@ -1,6 +1,6 @@
 {
 	"jsxSingleQuote": true,
-	"printWidth": 120,
+	"printWidth": 100,
 	"semi": false,
 	"singleQuote": true,
 	"tabWidth": 4,
