+++ package.json
@@ -3,26 +3,22 @@
 	"version": "1.0.0",
 	"private": true,
 	"dependencies": {
-		"@testing-library/jest-dom": "^5.14.1",
-		"@testing-library/react": "^12.1.2",
-		"@testing-library/user-event": "^7.2.1",
-		"@types/react-dom": "^17.0.9",
+		"@types/react-dom": "^17.0.11",
 		"@types/react-helmet": "^6.1.4",
-		"@types/react-router-dom": "^5.3.1",
+		"@types/react-router-dom": "^5.3.2",
 		"iframe-resizer-react": "^1.1.0",
 		"maxmsp-gui": "^1.0.3",
 		"p5": "^1.4.0",
 		"react": "^17.0.2",
 		"react-dom": "^17.0.2",
-		"react-ga": "^3.3.0",
 		"react-helmet": "^6.1.0",
-		"react-markdown": "^5.0.3",
-		"react-p5-wrapper": "^2.3.0",
-		"react-router-dom": "^5.3.0",
+		"react-markdown": "^7.1.1",
+		"react-p5-wrapper": "^3.0.2",
+		"react-router-dom": "^6.0.2",
 		"react-scripts": "^4.0.3",
-		"react-syntax-highlighter": "^15.4.3",
-		"sass": "^1.32.8",
-		"typescript": "^4.4.4"
+		"react-syntax-highlighter": "^15.4.5",
+		"sass": "^1.44.0",
+		"typescript": "^4.5.2"
 	},
 	"scripts": {
 		"start": "react-scripts start",
