--- public/config/music.json
@@ -1,32 +1,0 @@
-[
-	{
-		"title": "terracotta",
-		"src": "music-terracotta.png",
-		"released": true,
-		"link": "https://juliaset.bandcamp.com/album/terracotta"
-	},
-	{
-		"title": "morass / cairn",
-		"src": "music-morass.png",
-		"released": true,
-		"link": "https://juliaset.bandcamp.com/album/morass-cairn"
-	},
-	{
-		"link": "",
-		"released": false,
-		"src": "music-chroma.png",
-		"title": "Chroma / Weaving"
-	},
-	{
-		"link": "https://bois.bandcamp.com/album/noble",
-		"released": true,
-		"src": "music-noble.png",
-		"title": "Noble / Seed"
-	},
-	{
-		"link": "https://bois.bandcamp.com/album/bach-fractal",
-		"released": true,
-		"src": "music-bach.png",
-		"title": "bach.fractal"
-	}
-]
