--- public/config/scores.json
@@ -1,32 +1,0 @@
-[
-	{
-		"file": "Chroma_by_Lewis_Wolstanholme.pdf",
-		"instrumentation": "for live electronics and surround sound speakers",
-		"title": "Chroma",
-		"year": "2019"
-	},
-	{
-		"file": "Weaving_by_Lewis_Wolstanholme.pdf",
-		"instrumentation": "for three microtonal synthesisers",
-		"title": "Weaving",
-		"year": "2018 - 2019"
-	},
-	{
-		"file": "Quotient_by_Lewis_Wolstanholme.pdf",
-		"instrumentation": "for string trio",
-		"title": "ℚuotient",
-		"year": "2019"
-	},
-	{
-		"file": "Coprime_by_Lewis_Wolstanholme.pdf",
-		"instrumentation": "for piano trio",
-		"title": "Coprime",
-		"year": "2019"
-	},
-	{
-		"file": "Lorem_Ipsum_by_Lewis_Wolstanholme.pdf",
-		"instrumentation": "for microtonal piano",
-		"title": "Lorem Ipsum",
-		"year": "2018"
-	}
-]
