--- public/config/videos.json
@@ -1,26 +1,0 @@
-[
-	{
-		"hash": "LpO2JyVuqN4",
-		"title": "Terracotta by Julia Set"
-	},
-	{
-		"name": "Cairn by Julia Set",
-		"hash": "dVkmS5ku2jc"
-	},
-	{
-		"hash": "20eLg6skiK4",
-		"title": "Chroma by Lewis Wolf"
-	},
-	{
-		"hash": "eIMS1Q_vu_M",
-		"title": "Sabar by Julia Set"
-	},
-	{
-		"hash": "iXguBng4cJg",
-		"title": "Marble by Julia Set"
-	},
-	{
-		"hash": "sZYNHRozFBc",
-		"title": "Bach.Fractal by .bois"
-	}
-]
