--- src/components/info.tsx
@@ -1,10 +1,0 @@
-// dependencies
-import React from 'react'
-
-export default function Info(): JSX.Element {
-	return (
-		<div className='header' id='info-header'>
-			<p>lewiswolstanholme@gmail.com</p>
-		</div>
-	)
-}
