+++ src/react-app-env.d.ts
@@ -1,3 +1,4 @@
 /// <reference types="react-scripts" />
 
+declare module 'maxmsp-gui'
 declare module 'react-p5-wrapper'
