--- src/scss/code.scss
@@ -1,136 +1,0 @@
-#code-header .maxstuff {
-	height: fit-content;
-	width: 100%;
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: center;
-	justify-self: center;
-	align-self: self-end;
-
-	> div {
-		margin: 0 5px;
-	}
-
-	@media (max-width: $ipad) {
-		align-self: center;
-
-		> div {
-			&:first-of-type {
-				margin: 5px 5px 5px 0;
-			}
-			&:last-of-type {
-				margin: 5px 0 5px 5px;
-			}
-		}
-	}
-
-	@media (max-width: 375px) {
-		> div {
-			&:first-of-type {
-				margin: 5px 0;
-			}
-			&:last-of-type {
-				margin: 5px 0;
-			}
-		}
-	}
-}
-
-.code .repository {
-	iframe {
-		width: 100%;
-		border: 0;
-	}
-
-	> img {
-		height: auto;
-		width: 80%;
-		min-width: 600px;
-		max-width: 820px;
-		display: block;
-		margin: 0 auto 20px auto;
-		user-select: none;
-
-		@media (max-width: $ipad) {
-			width: 100%;
-			min-width: unset;
-			max-width: 600px;
-		}
-	}
-
-	.readme {
-		@media (max-width: $ipad) {
-			width: 100%;
-		}
-		width: 80%;
-		margin: 0 auto;
-		text-align: justify;
-		cursor: default;
-
-		pre {
-			overflow: scroll;
-			background: unset;
-		}
-
-		a {
-			color: black;
-			text-decoration: underline;
-		}
-
-		blockquote {
-			margin-left: 0;
-		}
-
-		h1 {
-			text-align: left;
-			margin: 0 0 20px 0;
-		}
-
-		h2 {
-			text-align: left;
-			font-size: 22px;
-		}
-
-		ul {
-			padding-left: 16px;
-		}
-
-		ol {
-			padding-left: 16px;
-		}
-
-		table {
-			border-collapse: collapse;
-			margin: 0 auto;
-			font-size: inherit;
-			overflow: scroll;
-
-			th {
-				width: 100px;
-				padding-bottom: 4px;
-				border-left: 2px solid #000000;
-
-				&:first-child {
-					border-left: none;
-				}
-			}
-
-			td {
-				padding-top: 4px;
-				border-left: 2px solid #000000;
-
-				&:first-child {
-					border-left: none;
-				}
-			}
-
-			thead {
-				border-bottom: 2px solid #000000;
-			}
-		}
-
-		p:last-of-type {
-			margin-bottom: 0;
-		}
-	}
-}
