--- src/scss/info.scss
@@ -1,11 +1,0 @@
-#info-header p {
-	margin: 0;
-	line-height: 24px;
-	vertical-align: middle;
-	justify-self: center;
-	align-self: self-end;
-
-	@media (max-width: $ipad) {
-		align-self: center;
-	}
-}
