--- src/scss/music.scss
@@ -1,36 +1,0 @@
-.music .albumcover {
-	height: 100%;
-	width: 100%;
-	position: relative;
-
-	&:focus {
-		outline: #b4b4b4 solid 5px;
-	}
-
-	img {
-		height: 100%;
-		width: 100%;
-		outline: 0;
-		user-select: none;
-	}
-
-	div {
-		height: 10%;
-		width: 100%;
-		position: absolute;
-		padding: 45% 0;
-		top: 0;
-		left: 0;
-		text-align: center;
-		color: white;
-		background: rgba(0, 0, 0, 0.6);
-		opacity: 0;
-		transition: opacity 0.5s;
-	}
-
-	&:hover {
-		div {
-			opacity: 1;
-		}
-	}
-}
