--- src/scss/scores.scss
@@ -1,63 +1,0 @@
-.scores .score-cover {
-	position: relative;
-	background: rgb(34, 31, 32);
-	cursor: pointer;
-
-	&:focus {
-		outline: #b4b4b4 solid 5px;
-	}
-
-	svg {
-		height: 100%;
-		width: 100%;
-		position: absolute;
-		top: 0;
-		left: 0;
-		outline: 0;
-	}
-
-	div {
-		position: relative;
-		outline: 0;
-		margin: 39px;
-
-		p {
-			position: absolute;
-			margin: 0;
-			padding: 0;
-			font-family: 'Helvetica';
-			font-weight: bold;
-			vertical-align: center;
-			color: white;
-			user-select: none;
-
-			&:nth-of-type(1) {
-				top: 0;
-				left: 0;
-				vertical-align: top;
-				font-size: 21px;
-				line-height: 21px;
-			}
-
-			&:nth-of-type(2) {
-				top: 0;
-				bottom: 0;
-				left: 0;
-				margin: auto;
-				text-align: justify;
-				vertical-align: center;
-				font-size: 12px;
-				line-height: 12px;
-				height: 25px;
-			}
-
-			&:nth-of-type(3) {
-				right: 0;
-				bottom: 0;
-				vertical-align: bottom;
-				font-size: 11px;
-				line-height: 11px;
-			}
-		}
-	}
-}
