--- src/syntax/index.js
@@ -1,1 +1,0 @@
-export { default as coy } from './coy';
